Edge case handling for superpwa_multisite_activation_status()
I had added this in the hope that it would be more efficient this way. But database option superpwa_version will still be present and that needs to be cleaned during uninstall.

get_option( 'superpwa_settings' ) can be big since its all the settings. Deleting this also saves that database call for each iteration of superpwa_multisite_network_deactivator().

It might be more efficient and cleaner this way.